Updated version of Jubi's patch.
The purpose of this patch is to support and test most or all undefined
behaviors for the supported subset of LLVM.

Relative to pull req #34, this one fixes up a few missing CamelCases
and also reduces the bitwidths used in several test cases.  The entire
"make check" now completes in about 2s on my machine, using a recent
STP. My judgement is that the bug-finding power of these test cases
is probably reduced very little or not at all by the reduced
bitwidths.