Added a new flag -souper-debug-level to control the verbose level of debug information
* Added one more cmdline option -souper-debug-level which
controls the verbose level of debug message. The larger
the number is, the more verbose message is generated.
The default level is 1.

* When the souper-debug-level is larger than 1, then dump all
replacements that collected by souper before they are processed
by KLEE. The existing debug-souper behavior will not be affected
if we do not pass a level higher than 1.