Fixed divide-by-zero crash for issue 84
We can't pass a constant zero as a divisor when building KLEE expressions,
because KLEE invokes llvm::APInt's div functions, which crash on zero value
divisors.

To fix this, when we build a div/rem expression, we fist check if the second
oprand is klee constant 0. If it is, then it definitely causes UB, and we
simply skip building the corresponding KLEE expression and just return a
constant zero.