Added preconditions to the UBInstrs collected through blockpc
In the original blockpc commit, we actually didn't consider
UB instructions on blockpc(s), because EB.getUBInstCondition was
invoked before blockpcs were processed. The problem was exposed
by the commit 22989c0 ("Handle instruction UB on the RHS correctly"),
where the order of getUBInstCondition and getBlockPCs were reversed.

Consequently, souper incorrectly produced the following smt query (simplified)
for the bug-induction opt file:

  (and (blockpcs ...) (and (ub-expr on "subnsw 2147483647:i32, %1")
                           (ub-expr on "subnsw 2147483648:i32, %1")))

This is wrong because those two instructions are conditional: they can
be considered only if the relevant blockpcs are true. So, the correct
form should be:

  (and (blockpcs ...) (and (relevant blockpc => ub-expr on "subnsw 2147483647:i32, %1")
                            ...))

where "=>" stands for "imply".

This patch fixed this issue (issue #154)