Handle llvm.assume
llvm.assume is turned into a path condition. We need
to be careful here not to add the pc to a replacement
in case the pc occurs after the instruction we
are looking to replace. Intuitively, if the pc
occurs after the instruction, then the instruction
can be part of the expression forming the assumption,
and in that case we would replace the assumption
by llvm.assume(true), which is not very useful.