Skip blockpc(s) where the block's phis are not relevant
When we generate relevant blockpc(s), we have to filter out
those where the block's phi instructions are not relevant to
the replacement candicate. Let's look at the following (simplified)
example:

; <label>:10                                      ; preds = %3, %7
  %11 = phi i32 [ %9, %7 ], [ %6, %3 ]
  store i32 %11, i32* @x5, align 4, !tbaa !1
  br i1 false, label %13, label %12

; <label>:12                                      ; preds = %10
  store i32 0, i32* @x3, align 4, !tbaa !1
  br label %13

; <label>:13                                      ; preds = %._crit_edge, %10, %12
  %14 = phi i32 [ %11, %10 ], [ %.pre, %._crit_edge ], [ %11, %12 ]
  store i32 %14, i32* @x4, align 4, !tbaa !1
  %16 = icmp eq i32 %1, 0

Previously, when we generate replacement candidate for instruction %16,
we will generate blockpc coming from [%11, %12], e.g.,

...
%10:i1 = slt %1, %9
blockpc %3 2 %10 0:i1
%11:i1 = eq 0:i32, %1
cand %11 0:i1

because %10 relates to %11 (they have %1 in common). However, it's bad
because the phi node is missing. Consequently, when we generate KLEE expressions,
blockpc %3 won't be touched because it can't be reached from any phi instruction,
and hence we will generate insufficient queries and mis-optimize the given code.

This patch fixed this issue (#166).