add a new option for dynamic profiling of all sites that souper might optimize, not just the ones that it has been able to optimize. this is useful because it permits us to find the highest profile sites and then run expensive synthesis only on those, instead of doing synthesis first.
this profile-all mode needs to run last: it can't run as a peephole
pass.  unfortunately it isn't possible to make a single pass that is
scheduled either as a peephole or running last, so this patch compiles
two versions of the pass and then sclang chooses the right one

also, it turns out that running profile code from an LLVM function
pass was incorrect and only happened to work: it is illegal to add new
functions from a function pass. so turn the function pass into a
module pass.

finally, fix a different bug in the profile instrumentation where we
had failed to put the atomic increment below any phi nodes that happen
to be there.