Support for LLVM IR funnel shift left/right intrinsic.
LangRef: https://releases.llvm.org/7.0.0/docs/LangRef.html#llvm-fshl-intrinsic

```
The ‘llvm.fshl’/‘llvm.fshr’ family of intrinsic functions performs
a funnel shift left/right: the first two values are concatenated
as { %a : %b } (%a is the most significant bits of the wide value),
the combined value is shifted left/right,
and the most/least significant bits are extracted to produce a result
that is the same size as the original arguments.
<...>
For vector types, the operation occurs for each element of the vector.
The shift argument is treated as an unsigned amount modulo
the element size of the arguments.
```

Observations:
* If first and second arguments are the same, it's just a normal rotate.
* If it's ‘llvm.fshl’, and second argument is `0`, it's just `shl`.
* If it's ‘llvm.fshr’, and first argument is `0`, it's just `lshr`.
* If it's ‘llvm.fshl’, and third argument is `0`, it's just first argument.
* If it's ‘llvm.fshr’, and third argument is `0`, it's just second argument.

Doubts:
* Do we really want to be this aggressive in synthesizing `fshl`/`fshr`?
  Currently, i tuned `Inst::getCost()` to synthesize it in all the cases
  covered in newly-added `test/Pass/` tests.
* I did not check exhaustive synthesis (no tests).
* First real patch to souper, i'm sure i have missed things,
  or failed at test coverage.

Bugs:
* If the shift amount is constant, even though
  `InstSynthesis::createCleanInst()` does reduce the value of that
  constant (makes it modulo bitwidth), it does not want to use this
  replacement, since the cost is identical.

Bonus:
* Added some docs to the `test/*/README.md` with my understanding of
  what test should be put where.