cmake: prevent picking up random LLVM installation
So I just faced this situation. I had a Debug build of LLVM in
third_party/ but I was only configuring a Release build of souper.
cmake's find_program would fail to find the Release/ build in
third_party and would pick system-level LLVM. It's better to error out
early in configure step in such cases.