Fix: Always creates new copy of reserved consts when doing wiring in multiple instruction synthesis.
Before the fix, C1 in below points the same reservedconst var:

Inst1 C1  Inst2 C1
  \  /      \  /
  mul        mul

This is wrong. The cache in getInstCopy() when filling in constants after first query might hit the wrong value. This patch will gurantee C1 and C2 always points to different objects.

Also this patch fixes a potential error, when creating an unary operator with a reserved inst as its argument, always call getReservedInst() for a new value. This also keeps the cache from breaking things.